Split Outline and Intra build targets
