+++ outline/apple/tun2socks.go
@@ -23,8 +23,8 @@ import (
 	"github.com/Jigsaw-Code/outline-go-tun2socks/tunnel"
 )
 
-// AppleTunnel embeds the tun2socks.Tunnel interface so it gets exported by gobind.
-type AppleTunnel interface {
+// OutlineTunnel embeds the tun2socks.Tunnel interface so it gets exported by gobind.
+type OutlineTunnel interface {
 	tunnel.OutlineTunnel
 }
 
@@ -39,14 +39,14 @@ func init() {
 	debug.SetGCPercent(10)
 	ticker := time.NewTicker(time.Minute * 1)
 	go func() {
-		for _ = range ticker.C {
+		for range ticker.C {
 			debug.FreeOSMemory()
 		}
 	}()
 }
 
 // ConnectSocksTunnel reads packets from a TUN device and routes it to a SOCKS server. Returns an
-// AppleTunnel instance that should be used to input packets to the tunnel.
+// OutlineTunnel instance that should be used to input packets to the tunnel.
 //
 // `tunWriter` is used to output packets to the TUN (VPN).
 // `host` is the IP address of the SOCKS proxy server.
@@ -54,9 +54,9 @@ func init() {
 // `isUDPEnabled` indicates whether the tunnel and/or network enable UDP proxying.
 //
 // Sets an error if the tunnel fails to connect.
-func ConnectSocksTunnel(tunWriter TunWriter, host string, port int, isUDPEnabled bool) (AppleTunnel, error) {
+func ConnectSocksTunnel(tunWriter TunWriter, host string, port int, isUDPEnabled bool) (tunnel.OutlineTunnel, error) {
 	if tunWriter == nil || host == "" || port <= 0 || port > 65535 {
 		return nil, errors.New("Must provide a TunWriter, a valid SOCKS proxy host and port")
 	}
-	return tunnel.NewTunnel(host, uint16(port), isUDPEnabled, tunWriter)
+	return tunnel.NewOutlineTunnel(host, uint16(port), isUDPEnabled, tunWriter)
 }
